+++ src/components/core/Drawer.vue
@@ -118,10 +118,10 @@ export default {
     responsive: false
   }),
   computed: {
-    ...mapState('app', ['image', 'color']),
+    ...mapState('drawerapp', ['image', 'color']),
     inputValue: {
       get () {
-        return this.$store.state.app.drawer
+        return this.$store.state.drawerapp.drawer
       },
       set (val) {
         this.setDrawer(val)
@@ -139,7 +139,8 @@ export default {
     window.removeEventListener('resize', this.onResponsiveInverted)
   },
   methods: {
-    ...mapMutations('app', ['setDrawer', 'toggleDrawer']),
+    // imports and spreads the mutations into their own variables, easier than importing one by one
+    ...mapMutations('drawerapp', ['setDrawer', 'toggleDrawer']),
     onResponsiveInverted () {
       if (window.innerWidth < 991) {
         this.responsive = true
