--- src/store/modules/app/mutations.js
@@ -1,8 +1,0 @@
-import { set, toggle } from '@/utils/vuex'
-
-export default {
-  setDrawer: set('drawer'),
-  setImage: set('image'),
-  setColor: set('color'),
-  toggleDrawer: toggle('drawer')
-}
