--- src/utils/vuex.js
@@ -1,3 +1,0 @@
-export const set = property => (state, payload) => (state[property] = payload)
-
-export const toggle = property => state => (state[property] = !state[property])
